requirements.txt
addict==2.4.0
aiofiles==23.2.1
albucore==0.0.21
albumentations==1.4.22
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
Arpeggio==2.0.2
audioop-lts==0.2.1
bitarray==2.9.3
bitmath==1.3.3.1
bitstring==4.2.3
blessed==1.17.12
bpylist2==4.1.1
certifi==2024.8.30
cffi==1.17.1
cgmetadata==0.2.0
charset-normalizer==3.4.0
click==8.1.7
Cython==3.0.11
eval_type_backport==0.2.0
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
fsspec==2024.10.0
gradio==5.8.0
gradio_client==1.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.36.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jpeg4py==0.1.4
lazy_loader==0.4
loguru==0.7.3
mac-alias==2.2.2
macholib==1.16.3
makelive==0.6.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.14.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
objexplore==1.6.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.12
osxmetadata==1.4.0
osxphotos==0.69.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathvalidate==3.2.1
photoscript==0.4.0
pillow==11.0.0
pillow_heif==0.21.0
prompt_toolkit==3.0.48
ptpython==3.0.29
py-applescript==1.0.3
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyobjc-core==10.3.2
pyobjc-framework-AppleScriptKit==10.3.2
pyobjc-framework-AppleScriptObjC==10.3.2
pyobjc-framework-AVFoundation==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Contacts==10.3.2
pyobjc-framework-CoreAudio==10.3.2
pyobjc-framework-CoreLocation==10.3.2
pyobjc-framework-CoreMedia==10.3.2
pyobjc-framework-CoreML==10.3.2
pyobjc-framework-CoreServices==10.3.2
pyobjc-framework-FSEvents==10.3.2
pyobjc-framework-Metal==10.3.2
pyobjc-framework-Photos==10.3.2
pyobjc-framework-Quartz==10.3.2
pyobjc-framework-ScriptingBridge==10.3.2
pyobjc-framework-UniformTypeIdentifiers==9.2
pyobjc-framework-Vision==10.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytimeparse2==1.7.1
pytz==2024.2
PyYAML==6.0.2
rawpy @ file:///Users/davidmerrick/code/rawpy
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-theme-manager==0.11.0
ruff==0.8.2
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.6.0
shellingham==1.5.4
shortuuid==1.0.13
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
stringzilla==3.11.0
strpdatetime==0.4.0
sympy==1.13.1
tenacity==8.5.0
textX==4.1.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.12
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0.dev20241206
torchaudio==2.5.0.dev20241206
torchvision==0.20.0.dev20241206
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
utitools==0.2.0
uvicorn==0.32.1
wcwidth==0.2.13
websockets==14.1
wheel==0.45.1
wrapt==1.17.0
wurlitzer==3.1.1
xattr==1.1.0
xdg-base-dirs==6.0.2
